Split out getJetpackProductDisplayName
