+++ client/lib/products-values/get-jetpack-product-display-name.js
@@ -1,0 +1,20 @@
+/**
+ * Internal dependencies
+ */
+import { assertValidProduct } from 'lib/products-values/utils/assert-valid-product';
+import { formatProduct } from 'lib/products-values/format-product';
+import { getJetpackProductsDisplayNames } from 'lib/products-values/translations';
+
+/**
+ * Get Jetpack product display name based on the product purchase object.
+ *
+ * @param   product {object}             Product purchase object
+ * @returns         {string|object} Product display name
+ */
+export function getJetpackProductDisplayName( product ) {
+	product = formatProduct( product );
+	assertValidProduct( product );
+	const jetpackProductsDisplayNames = getJetpackProductsDisplayNames();
+
+	return jetpackProductsDisplayNames?.[ product.productSlug ];
+}
